optimized engine loop replaced scoreboard sorting algorithm with base2^16 single-pass Radix sort

+++ src/engine.h
@@ -40,7 +40,7 @@ class Engine : public EventReceiver {
     void refreshScoreBoard();
     void issueOptimalTransfers();
     void setSpeedScale();
-    int calculateScore(File *, Pointer<Race>, FileList *, SiteRace *, FileList *, SiteRace *, int, bool *, bool) const;
+    unsigned short calculateScore(File *, Pointer<Race>, FileList *, SiteRace *, FileList *, SiteRace *, int, bool *, bool) const;
     void checkIfRaceComplete(SiteLogic *, Pointer<Race>);
     void raceComplete(Pointer<Race>);
     void transferJobComplete(Pointer<TransferJob>);
